Edit top bar, move liked songs to new folder
