+++ IT008_DoAnCuoiKi/Data/API/RSearch.cs
@@ -20,7 +20,7 @@ namespace IT008_DoAnCuoiKi.Data.API
         {
             var client = new RestClient("https://api.spotify.com/v1/search");
             client.AddDefaultHeader("Authorization", $"Bearer {TokenString.AccessToken}");
-            var request = new RestRequest($"?q={search_word}&type={type}", Method.Get);
+            var request = new RestRequest($"?q={search_word}&type={type}&limit=50", Method.Get);
             try
             {
                 var response = client.Execute(request);
