+++ IT008_DoAnCuoiKi/IT008_DoAnCuoiKi.csproj
@@ -125,6 +125,7 @@
     <Compile Include="Data\Models\MSearch.cs" />
     <Compile Include="Data\Models\MSongs.cs" />
     <Compile Include="Data\Models\MProfile.cs" />
+    <Compile Include="Data\Models\MTracks.cs" />
     <Compile Include="MainWindow.xaml.cs">
       <DependentUpon>MainWindow.xaml</DependentUpon>
       <SubType>Code</SubType>
@@ -137,7 +138,7 @@
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
-    <Page Include="Pages\LikedSongs.xaml">
+    <Page Include="Pages\LikedSongsPage\LikedSongs.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
@@ -179,7 +180,7 @@
     <Compile Include="Pages\Home.xaml.cs">
       <DependentUpon>Home.xaml</DependentUpon>
     </Compile>
-    <Compile Include="Pages\LikedSongs.xaml.cs">
+    <Compile Include="Pages\LikedSongsPage\LikedSongs.xaml.cs">
       <DependentUpon>LikedSongs.xaml</DependentUpon>
     </Compile>
     <Compile Include="Pages\Queue.xaml.cs">
