+++ IT008_DoAnCuoiKi/MainWindow.xaml
@@ -53,8 +53,22 @@
                 </Grid.RowDefinitions>
 
                 <ListBox SelectionMode="Single" x:Name="sidebar" SelectionChanged="sidebar_SelectionChanged" Background="Transparent" BorderThickness="0" Grid.Row="1" Grid.RowSpan="10" ScrollViewer.VerticalScrollBarVisibility="Hidden" ScrollViewer.HorizontalScrollBarVisibility="Hidden">
-                    <local:NavButton x:Name="nav_home" ToolTip="Home" Margin="0,20,0,0" Padding="6" Navlink="/Pages/Home.xaml" Icon="M8.354 1.146a.5.5 0 0 0-.708 0l-6 6A.5.5 0 0 0 1.5 7.5v7a.5.5 0 0 0 .5.5h4.5a.5.5 0 0 0 .5-.5v-4h2v4a.5.5 0 0 0 .5.5H14a.5.5 0 0 0 .5-.5v-7a.5.5 0 0 0-.146-.354L13 5.793V2.5a.5.5 0 0 0-.5-.5h-1a.5.5 0 0 0-.5.5v1.293L8.354 1.146ZM2.5 14V7.707l5.5-5.5 5.5 5.5V14H10v-4a.5.5 0 0 0-.5-.5h-3a.5.5 0 0 0-.5.5v4H2.5Z" />
-                    <local:NavButton ToolTip="Search" x:Name="nav_search" Margin="0,20,0,0" Padding="6" Navlink="/Pages/Search.xaml" Icon="M11.742 10.344a6.5 6.5 0 1 0-1.397 1.398h-.001c.03.04.062.078.098.115l3.85 3.85a1 1 0 0 0 1.415-1.414l-3.85-3.85a1.007 1.007 0 0 0-.115-.1zM12 6.5a5.5 5.5 0 1 1-11 0 5.5 5.5 0 0 1 11 0z" />
+                    <local:NavButton
+                        x:Name="nav_home"
+                        ToolTip="Home"
+                        Margin="0,20,0,0"
+                        Padding="6"
+                        Navlink="/Pages/Home.xaml"
+                        Icon="M8.354 1.146a.5.5 0 0 0-.708 0l-6 6A.5.5 0 0 0 1.5 7.5v7a.5.5 0 0 0 .5.5h4.5a.5.5 0 0 0 .5-.5v-4h2v4a.5.5 0 0 0 .5.5H14a.5.5 0 0 0 .5-.5v-7a.5.5 0 0 0-.146-.354L13 5.793V2.5a.5.5 0 0 0-.5-.5h-1a.5.5 0 0 0-.5.5v1.293L8.354 1.146ZM2.5 14V7.707l5.5-5.5 5.5 5.5V14H10v-4a.5.5 0 0 0-.5-.5h-3a.5.5 0 0 0-.5.5v4H2.5Z"
+                        />
+                    <local:NavButton
+                        ToolTip="Search"
+                        x:Name="nav_search"
+                        Margin="0,20,0,0"
+                        Padding="6"
+                        Navlink="/Pages/Search.xaml"
+                        Icon="M11.742 10.344a6.5 6.5 0 1 0-1.397 1.398h-.001c.03.04.062.078.098.115l3.85 3.85a1 1 0 0 0 1.415-1.414l-3.85-3.85a1.007 1.007 0 0 0-.115-.1zM12 6.5a5.5 5.5 0 1 1-11 0 5.5 5.5 0 0 1 11 0z"
+                        />
                     <local:NavButton
                         ToolTip="Your Library"
                         Margin="0,20,0,0"
@@ -62,11 +76,23 @@
                         Navlink="/Pages/YourLibrary.xaml"
                         x:Name ="nav_your_lib"
                         Icon="M2.5 3.5a.5.5 0 0 1 0-1h11a.5.5 0 0 1 0 1h-11zm2-2a.5.5 0 0 1 0-1h7a.5.5 0 0 1 0 1h-7zM0 13a1.5 1.5 0 0 0 1.5 1.5h13A1.5 1.5 0 0 0 16 13V6a1.5 1.5 0 0 0-1.5-1.5h-13A1.5 1.5 0 0 0 0 6v7zm1.5.5A.5.5 0 0 1 1 13V6a.5.5 0 0 1 .5-.5h13a.5.5 0 0 1 .5.5v7a.5.5 0 0 1-.5.5h-13z" />
-                    <local:NavButton ToolTip="Create Playlist" Margin="0,20,0,0" Padding="6" Navlink="/Pages/CreatePlaylist.xaml" Icon="M2 0a2 2 0 0 0-2 2v12a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2V2a2 2 0 0 0-2-2H2zm6.5 4.5v3h3a.5.5 0 0 1 0 1h-3v3a.5.5 0 0 1-1 0v-3h-3a.5.5 0 0 1 0-1h3v-3a.5.5 0 0 1 1 0z" />
-                    <local:NavButton ToolTip="Liked Songs" Margin="0, 20,0,0" Padding="6" Navlink="/Pages/LikedSongs.xaml" Icon="M5.32068 28.3355C5.32068 28.3355 20.5889 45.7575 25.742 51.6003C26.6008 52.6692 28.0322 53.2862 29.4636 53.2862C30.9904 53.2862 32.3264 52.6692 33.2806 51.6003C37.1931 47.1259 47.3083 35.6872 53.7019 28.3355C57.9007 23.5244 59.2367 16.9567 57.0418 11.0404C55.515 6.9535 52.2705 3.67325 48.0718 2.02904C43.873 0.384832 39.1017 0.534554 34.9983 2.43918L29.4636 4.99988L24.0243 2.43918C19.9209 0.534554 15.1496 0.384832 10.9508 2.02904C6.75208 3.67325 3.50758 6.9535 1.98076 11.0404C-0.214054 16.9567 1.02649 23.5244 5.32068 28.3355ZM29.9407 47.1422C29.8453 47.2702 29.6544 47.3445 29.4636 47.3445C29.3682 47.3445 29.1773 47.2702 29.0819 47.1422C26.0282 43.7613 9.51945 24.9418 9.51945 24.9418C6.56123 21.5608 5.70239 16.9449 7.22921 12.7863C8.18348 10.1285 10.2829 7.99518 13.0502 6.92536C15.8176 5.85644 18.8712 5.95355 21.5432 7.19306C25.2648 8.9144 28.3185 10.3236 28.3185 10.3236C29.0819 10.6829 29.9407 10.6829 30.7041 10.3236L37.4794 7.19306C40.0559 5.95355 43.205 5.85644 45.9724 6.92536C48.6443 7.99518 50.8391 10.1285 51.7934 12.7863C53.3202 16.9449 52.4614 21.5608 49.4077 24.9418C49.4077 24.9418 32.8989 43.7613 29.9407 47.1422Z" />
+                    <local:NavButton
+                        ToolTip="Create Playlist"
+                        Margin="0,20,0,0"
+                        Padding="6"
+                        Navlink="/Pages/CreatePlaylist.xaml"
+                        Icon="M2 0a2 2 0 0 0-2 2v12a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2V2a2 2 0 0 0-2-2H2zm6.5 4.5v3h3a.5.5 0 0 1 0 1h-3v3a.5.5 0 0 1-1 0v-3h-3a.5.5 0 0 1 0-1h3v-3a.5.5 0 0 1 1 0z"
+                        />
+                    <local:NavButton
+                        ToolTip="Liked Songs"
+                        Margin="0, 20,0,0"
+                        Padding="6"
+                        Navlink="/Pages/LikedSongsPage/LikedSongs.xaml"
+                        Icon="M5.32068 28.3355C5.32068 28.3355 20.5889 45.7575 25.742 51.6003C26.6008 52.6692 28.0322 53.2862 29.4636 53.2862C30.9904 53.2862 32.3264 52.6692 33.2806 51.6003C37.1931 47.1259 47.3083 35.6872 53.7019 28.3355C57.9007 23.5244 59.2367 16.9567 57.0418 11.0404C55.515 6.9535 52.2705 3.67325 48.0718 2.02904C43.873 0.384832 39.1017 0.534554 34.9983 2.43918L29.4636 4.99988L24.0243 2.43918C19.9209 0.534554 15.1496 0.384832 10.9508 2.02904C6.75208 3.67325 3.50758 6.9535 1.98076 11.0404C-0.214054 16.9567 1.02649 23.5244 5.32068 28.3355ZM29.9407 47.1422C29.8453 47.2702 29.6544 47.3445 29.4636 47.3445C29.3682 47.3445 29.1773 47.2702 29.0819 47.1422C26.0282 43.7613 9.51945 24.9418 9.51945 24.9418C6.56123 21.5608 5.70239 16.9449 7.22921 12.7863C8.18348 10.1285 10.2829 7.99518 13.0502 6.92536C15.8176 5.85644 18.8712 5.95355 21.5432 7.19306C25.2648 8.9144 28.3185 10.3236 28.3185 10.3236C29.0819 10.6829 29.9407 10.6829 30.7041 10.3236L37.4794 7.19306C40.0559 5.95355 43.205 5.85644 45.9724 6.92536C48.6443 7.99518 50.8391 10.1285 51.7934 12.7863C53.3202 16.9449 52.4614 21.5608 49.4077 24.9418C49.4077 24.9418 32.8989 43.7613 29.9407 47.1422Z"
+                        />
                 </ListBox>
             </Grid>
-            <Grid Grid.ColumnSpan="5" Grid.Column="1" Background="#000" Panel.ZIndex="9999">
+            <Grid Grid.ColumnSpan="5" Grid.Column="1" Background="Transparent" Panel.ZIndex="9999">
                 <Grid.ColumnDefinitions>
                     <ColumnDefinition></ColumnDefinition>
                     <ColumnDefinition></ColumnDefinition>
@@ -141,7 +167,7 @@
                     </WrapPanel>
                 </Grid>
             </Grid>
-            <Frame Grid.Row="0" Grid.Column="1" Grid.RowSpan="13" x:Name="navframe" NavigationUIVisibility="Hidden" Source="/Pages/YourLibrary.xaml" />
+            <Frame Grid.Row="0" Grid.Column="1" Grid.RowSpan="13" x:Name="navframe" NavigationUIVisibility="Hidden" Source="/Pages/Home.xaml" />
         </Grid>
     </Grid>
