+++ IT008_DoAnCuoiKi/Pages/YourLibrary.xaml
@@ -43,7 +43,7 @@
                         Padding="6"
                         Navlink="/PlayListThing/ViewAlbum.xaml"
                         Width="142" Height="52"
-                        Icon="M28.4045 0.866302C12.7545 0.866302 0.0628052 12.9256 0.0628052 27.7797C0.0628052 42.6429 12.7545 54.7024 28.4045 54.7024C44.0544 54.7024 56.7461 42.6429 56.7461 27.7797C56.7461 12.9256 44.0544 0.866302 28.4045 0.866302ZM28.4045 6.16555C41.0008 6.16555 51.116 15.8474 51.116 27.7797C51.116 39.7211 41.0008 49.4031 28.4045 49.4031C15.8082 49.4031 5.69297 39.7211 5.69297 27.7797C5.69297 15.8474 15.8082 6.16555 28.4045 6.16555ZM28.4045 17.7258C22.5835 17.7258 17.8121 22.2355 17.8121 27.7797C17.8121 33.333 22.5835 37.8428 28.4045 37.8428C34.2255 37.8428 38.9968 33.333 38.9968 27.7797C38.9968 22.2355 34.2255 17.7258 28.4045 17.7258ZM28.4045 23.0159C31.1718 23.0159 33.4621 25.1573 33.4621 27.7797C33.4621 30.4112 31.1718 32.5528 28.4045 32.5528C25.6371 32.5528 23.3469 30.4112 23.3469 27.7797C23.3469 25.1573 25.6371 23.0159 28.4045 23.0159Z"/>"
+                        Icon="M28.4045 0.866302C12.7545 0.866302 0.0628052 12.9256 0.0628052 27.7797C0.0628052 42.6429 12.7545 54.7024 28.4045 54.7024C44.0544 54.7024 56.7461 42.6429 56.7461 27.7797C56.7461 12.9256 44.0544 0.866302 28.4045 0.866302ZM28.4045 6.16555C41.0008 6.16555 51.116 15.8474 51.116 27.7797C51.116 39.7211 41.0008 49.4031 28.4045 49.4031C15.8082 49.4031 5.69297 39.7211 5.69297 27.7797C5.69297 15.8474 15.8082 6.16555 28.4045 6.16555ZM28.4045 17.7258C22.5835 17.7258 17.8121 22.2355 17.8121 27.7797C17.8121 33.333 22.5835 37.8428 28.4045 37.8428C34.2255 37.8428 38.9968 33.333 38.9968 27.7797C38.9968 22.2355 34.2255 17.7258 28.4045 17.7258ZM28.4045 23.0159C31.1718 23.0159 33.4621 25.1573 33.4621 27.7797C33.4621 30.4112 31.1718 32.5528 28.4045 32.5528C25.6371 32.5528 23.3469 30.4112 23.3469 27.7797C23.3469 25.1573 25.6371 23.0159 28.4045 23.0159Z"/>
         </ListBox>
         <Frame x:Name="PlayList_Page" Background="#171717" Grid.Row="1" Grid.ColumnSpan="8" Grid.RowSpan="5" NavigationUIVisibility="Hidden" Source="/PlayListThing/PlayListPage.xaml" />
     </Grid>
