+++ IT008_DoAnCuoiKi/PlayListThing/ViewAlbum.xaml
@@ -12,7 +12,7 @@
 					<TextBlock Height="79"></TextBlock>
 					<Path Fill="White" Stretch="Uniform" Data="M28.4045 0.866302C12.7545 0.866302 0.0628052 12.9256 0.0628052 27.7797C0.0628052 42.6429 12.7545 54.7024 28.4045 54.7024C44.0544 54.7024 56.7461 42.6429 56.7461 27.7797C56.7461 12.9256 44.0544 0.866302 28.4045 0.866302ZM28.4045 6.16555C41.0008 6.16555 51.116 15.8474 51.116 27.7797C51.116 39.7211 41.0008 49.4031 28.4045 49.4031C15.8082 49.4031 5.69297 39.7211 5.69297 27.7797C5.69297 15.8474 15.8082 6.16555 28.4045 6.16555ZM28.4045 17.7258C22.5835 17.7258 17.8121 22.2355 17.8121 27.7797C17.8121 33.333 22.5835 37.8428 28.4045 37.8428C34.2255 37.8428 38.9968 33.333 38.9968 27.7797C38.9968 22.2355 34.2255 17.7258 28.4045 17.7258ZM28.4045 23.0159C31.1718 23.0159 33.4621 25.1573 33.4621 27.7797C33.4621 30.4112 31.1718 32.5528 28.4045 32.5528C25.6371 32.5528 23.3469 30.4112 23.3469 27.7797C23.3469 25.1573 25.6371 23.0159 28.4045 23.0159Z" Height="81" Width="70"/>
 					<TextBlock></TextBlock>
-					<TextBlock Grid.Row="1" TextWrapping="Wrap" Text="Follow your first album" Foreground="White" FontSize="24" FontWeight="Bold" Height="67" FontFamily="Arial Black" TextAlignment="Center" Grid.ColumnSpan="2"/>
+					<TextBlock Grid.Row="1" x:Name="Test" TextWrapping="Wrap" Text="Follow your first album" Foreground="White" FontSize="24" FontWeight="Bold" Height="67" FontFamily="Arial Black" TextAlignment="Center" Grid.ColumnSpan="2"/>
 					<TextBlock Grid.Row="1" TextWrapping="Wrap" Text="Save albums by tapping the heart icons." Foreground="White" FontSize="15" FontWeight="Bold" Height="65" FontFamily="Arial Black" TextAlignment="Center" Grid.ColumnSpan="2"/>
 					<Button x:Name="FindArtist" Background="White" Content="Find albums" FontSize="15" FontFamily="Arial Black" Grid.Column="1" Margin="48,21,47,20" Grid.Row="1" Width="200" RenderTransformOrigin="0.5,0.5" Height="52">
 						<Button.Resources>
